fixes and additions
A few little bug fixes in error handling and the way data is appended to
the database.

Changed the way the auto status functionality works and expanded the
amount of things to select from.

Also added a new database to log the tweet ids of automatically posted
statuses. Now a function checks tweets mentioning my username with a
reply to tweet id that matches one in my autolog database. Sends
response if conditons met.